extract mustache utils
